Hike platform internal

+++ src/com/bsb/hike/platform/bridge/NonMessagingJavaScriptBridge.java
@@ -1,5 +1,6 @@
 package com.bsb.hike.platform.bridge;
 
+import com.bsb.hike.bots.BotUtils;
 import org.json.JSONException;
 import org.json.JSONObject;
 
@@ -697,7 +698,7 @@ public class NonMessagingJavaScriptBridge extends JavascriptBridge
 	}
 
 	/**
-	 * Platform Version 5
+	 * Platform Version 6
 	 * Call this function to delete an event from the list of events that are shared with the microapp.
 	 *
 	 * @param eventId: the event that will be deleted from the shared messages table.
@@ -714,7 +715,7 @@ public class NonMessagingJavaScriptBridge extends JavascriptBridge
 	}
 
 	/**
-	 * Platform Version 5
+	 * Platform Version 6
 	 * Call this function to delete all the events, be it shared data or normal event pertaining to a single message.
 	 *
 	 * @param messageHash
@@ -731,7 +732,7 @@ public class NonMessagingJavaScriptBridge extends JavascriptBridge
 	}
 
 	/**
-	 * Platform version 5
+	 * Platform version 6
 	 * Call this function to delete all the events for a particular microapp, be it shared data or normal event.
 	 */
 	@JavascriptInterface
@@ -741,7 +742,7 @@ public class NonMessagingJavaScriptBridge extends JavascriptBridge
 	}
 
 	/**
-	 * Platform Version 5
+	 * Platform Version 6
 	 * This function is made for the special Shared bot that has the information about some other bots as well, and acts as a channel for them.
 	 * Call this function to delete all the events for a particular microapp, be it shared data or normal event.
 	 *
@@ -755,11 +756,17 @@ public class NonMessagingJavaScriptBridge extends JavascriptBridge
 			Logger.e(TAG, "the events corresponding to the namespace can't be deleted as the namespace is " + namespace);
 			return;
 		}
+		NonMessagingBotMetadata metadata = new NonMessagingBotMetadata(mBotInfo.getMetadata());
+		if (!metadata.isSpecialBot())
+		{
+			Logger.e(TAG, "the bot is not a special bot and only special bot has the authority to call this function.");
+			return;
+		}
 		HikeConversationsDatabase.getInstance().deleteAllEventsForNamespace(namespace);
 	}
 
 	/**
-	 * Platform Version 5
+	 * Platform Version 6
 	 * Call this function to get all the shared messages data. The data is a stringified list that contains event id, message hash and the data.
 	 * <p/>
 	 * "name": name of the user interacting with. This gives name, and if the name isn't present , then the msisdn.
@@ -779,7 +786,7 @@ public class NonMessagingJavaScriptBridge extends JavascriptBridge
 	}
 
 	/**
-	 * Platform Version 5
+	 * Platform Version 6
 	 * This function is made for the special Shared bot that has the information about some other bots as well, and acts as a channel for them.
 	 * Call this function to get all the shared events data. The data is a stringified list that contains :
 	 * "name": name of the user interacting with. This gives name, and if the name isn't present , then the msisdn.
@@ -800,12 +807,18 @@ public class NonMessagingJavaScriptBridge extends JavascriptBridge
 			Logger.e(TAG, "can't return shared events as the namespace is " + namespace);
 			return;
 		}
+		NonMessagingBotMetadata metadata = new NonMessagingBotMetadata(mBotInfo.getMetadata());
+		if (!metadata.isSpecialBot())
+		{
+			Logger.e(TAG, "the bot is not a special bot and only special bot has the authority to call this function.");
+			return;
+		}
 		String messageData = HikeConversationsDatabase.getInstance().getMessageEventsForMicroapps(namespace, false);
 		callbackToJS(functionId, messageData);
 	}
 
 	/**
-	 * Platform Version 5
+	 * Platform Version 6
 	 * This function is made for the special Shared bot that has the information about some other bots as well, and acts as a channel for them.
 	 * Call this function to get all the event messages data. The data is a stringified list that contains:
 	 * "name": name of the user interacting with. This gives name, and if the name isn't present , then the msisdn.
@@ -827,12 +840,18 @@ public class NonMessagingJavaScriptBridge extends JavascriptBridge
 			Logger.e(TAG, "can't return all events as the namespace is " + namespace);
 			return;
 		}
+		NonMessagingBotMetadata metadata = new NonMessagingBotMetadata(mBotInfo.getMetadata());
+		if (!metadata.isSpecialBot())
+		{
+			Logger.e(TAG, "the bot is not a special bot and only special bot has the authority to call this function.");
+			return;
+		}
 		String messageData = HikeConversationsDatabase.getInstance().getMessageEventsForMicroapps(namespace, true);
 		callbackToJS(functionId, messageData);
 	}
 
 	/**
-	 * Platform Version 5
+	 * Platform Version 6
 	 * Call this function to get all the event messages data. The data is a stringified list that contains event id, message hash and the data.
 	 * <p/>
 	 * "name": name of the user interacting with. This gives name, and if the name isn't present , then the msisdn.
@@ -853,7 +872,7 @@ public class NonMessagingJavaScriptBridge extends JavascriptBridge
 	}
 
 	/**
-	 * Platform Version 5
+	 * Platform Version 6
 	 * Call this function to get all the event messages data. The data is a stringified list that contains:
 	 * "name": name of the user interacting with. This gives name, and if the name isn't present , then the msisdn.
 	 * "platformUid": the platform user id of the user interacting with.
@@ -873,7 +892,7 @@ public class NonMessagingJavaScriptBridge extends JavascriptBridge
 	}
 
 	/**
-	 * Platform Version 5
+	 * Platform Version 6
 	 * Call this function to send a shared message to the contacts of the user. This function when forwards the data, returns with the contact details of
 	 * the users it has sent the message to.
 	 * It will call JavaScript function "onContactChooserResult(int resultCode,JsonArray array)" This JSOnArray contains list of JSONObject where each JSONObject reflects one user. As of now
@@ -892,7 +911,7 @@ public class NonMessagingJavaScriptBridge extends JavascriptBridge
 	}
 
 	/**
-	 * Platform version 5
+	 * Platform version 6
 	 * Call this method to send a normal event.
 	 *
 	 * @param messageHash : the message hash that determines the uniqueness of the card message, to which the data is being sent.
@@ -906,7 +925,7 @@ public class NonMessagingJavaScriptBridge extends JavascriptBridge
 	}
 	
 	/**
-	* Platform Bridge Version 5
+	* Platform Bridge Version 6
 	* Call this method to post a status update without an image to timeline.
 	*
 	* @param status
@@ -1015,4 +1034,41 @@ public class NonMessagingJavaScriptBridge extends JavascriptBridge
 		
 		Utils.postStatusUpdate(status, mood, imageFilePath);
 	}
+
+	/**
+	 * Platform Version 6
+	 * This function is made for the special Shared bot that has the information about some other bots as well, and acts as a channel for them.
+	 * calling this method will forcefully block the full screen bot. The user won't see any messages in the bot after calling this.
+	 *
+	 * @param isBlocked : true to block the microapp false to unblock it.
+	 * @param msisdn : the msisdn of the bot to be blocked/unblocked
+	 */
+	@JavascriptInterface
+	public void blockBot(String isBlocked, String msisdn)
+	{
+		if (!BotUtils.isBot(msisdn))
+		{
+			return;
+		}
+		BotInfo botInfo = BotUtils.getBotInfoForBotMsisdn(msisdn);
+		NonMessagingBotMetadata metadata = new NonMessagingBotMetadata(botInfo.getMetadata());
+		if (!metadata.isSpecialBot())
+		{
+			Logger.e(TAG, "the bot is not a special bot and only special bot has the authority to call this function.");
+			return;
+		}
+		if (Boolean.valueOf(isBlocked))
+		{
+			botInfo.setBlocked(true);
+			HikeMessengerApp.getPubSub().publish(HikePubSub.BLOCK_USER, msisdn);
+		}
+
+		else
+		{
+			botInfo.setBlocked(false);
+			HikeMessengerApp.getPubSub().publish(HikePubSub.UNBLOCK_USER, msisdn);
+		}
+	}
+
+
 }
